Refactor main.py: Simplify code, improve validation, remove TODOs
